moved logging definitions to new file
